+++ app/src/main/java/com/example/adore/models/responses/CurrentUserResponse.kt
@@ -1,3 +1,3 @@
 package com.example.adore.models.responses
 
-data class CurrentUserResponse(val userId: Int)
+data class CurrentUserResponse(val userId: Long)
